+++ src/regex.ts
@@ -1,0 +1,5 @@
+import { variablePattern } from "./patterns";
+
+export const variableReg = new RegExp(variablePattern);
+
+export const matchVariableReg = new RegExp(`(\\s+|$\{)(${variablePattern})$`);